+++ mirx/src/coding/mod.rs
@@ -7,6 +7,6 @@ mod buffer;
 mod lz4;
 mod pixel;
 mod rle;
-pub use lz4::{Lz4, Lz4DecodePlan, Lz4Error};
+pub use lz4::{Lz4, Lz4DecodePlan, Lz4Encoder, Lz4Error};
 pub use pixel::{Pixel, PixelDecodePlan, PixelError};
 pub use rle::{Rle, RleDecodePlan, RleError};